refactor: improve structure with the use of modules
