+++ src/menu.js
@@ -1,0 +1,13 @@
+'use strict';
+
+/**
+ *
+ * @type {{pizza: string, hotdog: string, hamburger: string}}
+ */
+const menu = {
+  hamburger: 'Combo Hamburguesa',
+  hotdog: 'Combo Hot Dogs',
+  pizza: 'Combo Pizza',
+};
+
+module.exports = menu;
