+++ src/table.js
@@ -1,0 +1,9 @@
+'use strict';
+
+/**
+ *
+ * @type {*[]}
+ */
+const table = ['Mesa 1', 'Mesa 2', 'Mesa 3', 'Mesa 4', 'Mesa 5'];
+
+module.exports = table;
